+++ src/linter.ts
@@ -224,7 +224,7 @@ export const getJsLinter: getAsyncLinter<Linter.LintMessage[], string> = async (
 		text,
 		opt: Linter.LegacyConfig | null | undefined,
 	) => {
-		const config: Linter.LegacyConfig = {...conf, ...opt};
+		const config: Linter.LegacyConfig & {filename?: string} = {...conf, ...opt};
 		if (!('rules' in config)) {
 			let {extends: ex = []} = config;
 			if (!Array.isArray(ex)) {
@@ -236,7 +236,7 @@ export const getJsLinter: getAsyncLinter<Linter.LintMessage[], string> = async (
 			config.extends = ex;
 		}
 		linter.config = config;
-		return esLinter.verify(text, config)
+		return esLinter.verify(text, config, config.filename)
 			.filter(({ruleId, message}) => message !== `Definition for rule '${ruleId}' was not found.`);
 	};
 	linter.fixer = (code, rule): string => esLinter.verifyAndFix(